make free space double

+++ kernel/defs.h
@@ -63,7 +63,7 @@ void            ramdiskrw(struct buf*);
 void*           kalloc(void);
 void            kfree(void *);
 void            kinit(void);
-float           kfreespace(void);
+double          kfreespace(void);
 
 // log.c
 void            initlog(int, struct superblock*);
